remove search from content api
