+++ backend/tracim_backend/lib/core/content.py
@@ -2,7 +2,6 @@
 from contextlib import contextmanager
 import datetime
 import os
-import re
 import traceback
 import typing
 
@@ -17,7 +16,6 @@ from sqlalchemy import desc
 from sqlalchemy import func
 from sqlalchemy import or_
 from sqlalchemy.orm import Query
-from sqlalchemy.orm import joinedload
 from sqlalchemy.orm.attributes import QueryableAttribute
 from sqlalchemy.orm.attributes import get_history
 from sqlalchemy.orm.exc import NoResultFound
@@ -49,7 +47,6 @@ from tracim_backend.exceptions import UnavailablePreview
 from tracim_backend.exceptions import WorkspacesDoNotMatch
 from tracim_backend.lib.core.notifications import NotifierFactory
 from tracim_backend.lib.core.userworkspace import RoleApi
-from tracim_backend.lib.search.models import SimpleContentSearchResponse
 from tracim_backend.lib.search.search_factory import SearchFactory
 from tracim_backend.lib.utils.logger import logger
 from tracim_backend.lib.utils.sanitizer import HtmlSanitizer
@@ -127,10 +124,6 @@ class AddCopyRevisionsResult(object):
         self.new_content = new_content
         self.new_children_dict = new_children_dict  # dict key is original content id
         self.original_children_dict = original_children_dict  # dict key is original content id
-
-
-SEARCH_SEPARATORS = ",| "
-SEARCH_DEFAULT_RESULT_NB = 10
 
 
 class ContentApi(object):
@@ -2111,100 +2104,6 @@ class ContentApi(object):
         NotifierFactory.create(
             config=self._config, current_user=self._user, session=self._session
         ).notify_content_update(content)
-
-    def get_keywords(self, search_string, search_string_separators=None) -> typing.List[str]:
-        """
-        :param search_string: a list of coma-separated keywords
-        :return: a list of str (each keyword = 1 entry
-        """
-
-        search_string_separators = search_string_separators or SEARCH_SEPARATORS
-
-        keywords = []
-        if search_string:
-            keywords = [
-                keyword.strip() for keyword in re.split(search_string_separators, search_string)
-            ]
-
-        return keywords
-
-    def search(
-        self,
-        keywords: typing.List[str],
-        size: typing.Optional[int] = SEARCH_DEFAULT_RESULT_NB,
-        offset: typing.Optional[int] = None,
-        content_types: typing.Optional[typing.List[str]] = None,
-    ) -> SimpleContentSearchResponse:
-        query = self._search_query(keywords=keywords, content_types=content_types)
-        results = []
-        current_offset = 0
-        parsed_content_ids = []
-        for content in query:
-            if len(results) >= size:
-                break
-            if not self._show_deleted:
-                if self.get_deleted_parent_id(content):
-                    continue
-            if not self._show_archived:
-                if self.get_archived_parent_id(content):
-                    continue
-            if content.type == content_type_list.Comment.slug:
-                # INFO - G.M - 2019-06-13 -  filter by content_types of parent for comment
-                # if correct content_type, content is parent.
-                if content.parent.type in content_types:
-                    content = content.parent
-                else:
-                    continue
-            if content.content_id in parsed_content_ids:
-                # INFO - G.M - 2019-06-13 - avoid duplication of same content in result list
-                continue
-            if current_offset >= offset:
-                results.append(content)
-            parsed_content_ids.append(content.content_id)
-            current_offset += 1
-
-        content_in_context_list = []
-        for content in results:
-            content_in_context_list.append(self.get_content_in_context(content))
-        return SimpleContentSearchResponse(
-            content_list=content_in_context_list, total_hits=current_offset
-        )
-
-    def _search_query(
-        self, keywords: typing.List[str], content_types: typing.Optional[typing.List[str]] = None
-    ) -> Query:
-        """
-        :return: a sorted list of Content items
-        """
-
-        if len(keywords) <= 0:
-            return []
-
-        filter_group_label = list(
-            Content.label.ilike("%{}%".format(keyword)) for keyword in keywords
-        )
-        filter_group_filename = list(
-            Content.file_name.ilike("%{}%".format(keyword)) for keyword in keywords
-        )
-        filter_group_description = list(
-            Content.description.ilike("%{}%".format(keyword)) for keyword in keywords
-        )
-        title_keyworded_items = (
-            self.get_base_query(None)
-            .filter(or_(*(filter_group_label + filter_group_filename + filter_group_description)))
-            .options(joinedload("children_revisions"))
-            .options(joinedload("parent"))
-            .order_by(desc(Content.updated), desc(Content.revision_id), desc(Content.content_id))
-        )
-
-        # INFO - G.M - 2019-06-13 - we add comment to content_types checked
-        if content_types:
-            searched_content_types = set(content_types + [content_type_list.Comment.slug])
-            title_keyworded_items = title_keyworded_items.filter(
-                Content.type.in_(searched_content_types)
-            )
-
-        return title_keyworded_items
 
     def get_all_types(self) -> typing.List[ContentType]:
         labels = content_type_list.endpoint_allowed_types_slug()
